+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Upgrade to `clap` version 4.
 * Use `wild` for wildcard support on Windows.
 * Enable parallel volume analysis of multiple files.
+* Ensure input file is closed before renaming.
 
 ## 0.2.0